+++ datumaro/util/annotation_util.py
@@ -157,7 +157,7 @@ def segment_iou(a, b):
 
         a = _to_rle(a)
         b = _to_rle(b)
-    return float(mask_utils.iou(a, b, [not is_bbox]))
+    return float(mask_utils.iou(a, b, [not is_bbox]).item())
 
 
 def PDJ(a, b, eps=None, ratio=0.05, bbox=None):